Reinstate UIStarter.java file
Put UIStarter.java back with a shell of its code.  In ResultsTable.java, fix the class name and delete code that is not specific to the table.

This still will not, and is not intended, to compile.  This is being done piecemeal to be able to inspect the individual diffs in git.
