+++ src/main/org/tvrenamer/view/ResultsTable.java
@@ -16,8 +16,6 @@ import org.eclipse.swt.events.KeyAdapter;
 import org.eclipse.swt.events.KeyEvent;
 import org.eclipse.swt.events.SelectionAdapter;
 import org.eclipse.swt.events.SelectionEvent;
-import org.eclipse.swt.graphics.Image;
-import org.eclipse.swt.graphics.Rectangle;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.layout.RowLayout;
@@ -32,7 +30,6 @@ import org.eclipse.swt.widgets.Link;
 import org.eclipse.swt.widgets.Listener;
 import org.eclipse.swt.widgets.Menu;
 import org.eclipse.swt.widgets.MenuItem;
-import org.eclipse.swt.widgets.Monitor;
 import org.eclipse.swt.widgets.ProgressBar;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Table;
@@ -60,7 +57,6 @@ import org.tvrenamer.model.UserPreference;
 import org.tvrenamer.model.UserPreferences;
 import org.tvrenamer.model.util.Environment;
 
-import java.io.InputStream;
 import java.text.Collator;
 import java.util.LinkedList;
 import java.util.List;
@@ -68,11 +64,10 @@ import java.util.Locale;
 import java.util.Observable;
 import java.util.Observer;
 import java.util.Queue;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 
-public final class UIStarter implements Observer, AddEpisodeListener {
-    private static final Logger logger = Logger.getLogger(UIStarter.class.getName());
+public final class ResultsTable implements Observer, AddEpisodeListener {
+    private static final Logger logger = Logger.getLogger(ResultsTable.class.getName());
     private static final Collator COLLATOR = Collator.getInstance(Locale.getDefault());
 
     private static final int SELECTED_COLUMN = 0;
@@ -97,36 +92,6 @@ public final class UIStarter implements Observer, AddEpisodeListener {
     private boolean apiDeprecated = false;
 
     private final EpisodeDb episodeMap = new EpisodeDb();
-
-    private void init() {
-        // load preferences
-        prefs = UserPreferences.getInstance();
-        prefs.addObserver(this);
-
-        // Setup display and shell
-        GridLayout shellGridLayout = new GridLayout(3, false);
-        Display.setAppName(APPLICATION_NAME);
-        display = new Display();
-
-        shell = new Shell(display);
-
-        shell.setText(APPLICATION_NAME);
-        shell.setLayout(shellGridLayout);
-
-        // Setup the util class
-        UIUtils.setShell(shell);
-        UIUtils.checkDestinationDirectory(prefs);
-
-        // Add controls to main shell
-        setupMainWindow();
-        setupAddFilesDialog();
-        setupClearFilesButton();
-        setupMenuBar();
-
-        setupIcons();
-
-        shell.pack(true);
-    }
 
     private void setupUpdateStuff(final Composite parentComposite) {
         Link updatesAvailableLink = new Link(parentComposite, SWT.VERTICAL);
@@ -202,11 +167,6 @@ public final class UIStarter implements Observer, AddEpisodeListener {
                 uiCleanup();
             }
         });
-    }
-
-    private void uiCleanup() {
-        shell.dispose();
-        display.dispose();
     }
 
     private void makeMenuItem(Menu parent, String text, Listener listener, char shortcut) {
@@ -433,20 +393,6 @@ public final class UIStarter implements Observer, AddEpisodeListener {
         });
     }
 
-    private void setupIcons() {
-        try {
-            InputStream icon = getClass().getResourceAsStream(TVRENAMER_ICON_PATH);
-            if (icon != null) {
-                shell.setImage(new Image(display, icon));
-            } else {
-                shell.setImage(new Image(display, TVRENAMER_ICON_DIRECT_PATH));
-            }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
     Display getDisplay() {
         return display;
     }
@@ -457,46 +403,6 @@ public final class UIStarter implements Observer, AddEpisodeListener {
 
     TaskItem getTaskItem() {
         return taskItem;
-    }
-
-    private int launch() {
-        try {
-            // place the window in the centre of the primary monitor
-            Monitor primary = display.getPrimaryMonitor();
-            Rectangle bounds = primary.getBounds();
-            Rectangle rect = shell.getBounds();
-            int x = bounds.x + (bounds.width - rect.width) - 5;
-            int y = bounds.y + (bounds.height - rect.height) - 35;
-            shell.setLocation(x, y);
-
-            // Start the shell
-            shell.pack();
-            shell.open();
-            resultsTable.setFocus();
-
-            // Load the preload folder into the episode map, which will call
-            // us back with the list of files once they've been loaded.
-            episodeMap.subscribe(this);
-            episodeMap.preload();
-
-            while (!shell.isDisposed()) {
-                if (!display.readAndDispatch()) {
-                    display.sleep();
-                }
-            }
-            return 0;
-        } catch (Exception exception) {
-            showMessageBox(SWTMessageBoxType.ERROR, ERROR_LABEL, UNKNOWN_EXCEPTION, exception);
-            logger.log(Level.SEVERE, UNKNOWN_EXCEPTION, exception);
-            return 1;
-        }
-    }
-
-    public int run() {
-        init();
-        int rval = launch();
-        uiCleanup();
-        return rval;
     }
 
     /**
